requirements.txt
cffi~=1.16.0
defusedxml~=0.7.1
future~=1.0.0
pillow~=10.2.0
ipython~=8.26.0
jupyter~=1.0.0
pip~=24.0
attrs~=23.2.0
PySocks~=1.7.1
protobuf~=4.25.3
cryptography~=44.0.0
tornado~=6.4.2
Jinja2~=3.1.3
filelock~=3.16.1
requests~=2.31.0
ipywidgets~=8.1.3
notebook~=7.2.1
PyYAML~=6.0.1
pytest~=8.3.4
numpy~=1.26.4
setuptools~=72.0.0
pytz~=2023.4
overrides~=7.7.0
charset-normalizer~=3.3.2
numba~=0.59.1
regex~=2024.11.6
sympy~=1.13.1
mpmath~=1.3.0
decorator~=5.1.1
torch~=2.5.1
typing_extensions~=4.10.0
matplotlib~=3.8.3
tabulate~=0.9.0
tqdm~=4.66.2
scipy~=1.13.0rc1
psutil~=6.0.0
packaging~=23.2
urllib3~=2.2.1
fsspec~=2024.12.0
MarkupSafe~=2.1.5
certifi~=2024.8.30
networkx~=3.4.2
pandas~=2.1.4
idna~=3.6
chardet~=5.2.0
tiktoken~=0.8.0
tokenizers~=0.21.0
safetensors~=0.5.2
scikit-learn~=1.4.1.post1
transformers~=4.48.0
huggingface-hub~=0.27.1
Pygments~=2.18.0
sentencepiece~=0.2.0
nltk~=3.9.1
beautifulsoup4~=4.12.3
platformdirs~=4.2.2
Flask~=3.0.2
pdfplumber~=0.11.5
deep-translator~=1.11.4
python-dotenv~=0.21.1